requirements.txt
# コード品質とフォーマット
ruff

# 型チェック
mypy

# テスト
pytest

# Google Calendar API
google-api-python-client == 1.7.2
google-auth-oauthlib == 0.4.1

# Gemini API
google-generativeai

# .env support
python-dotenv
